ApplyTypeAnnotations: remove duplicate staticmethod
I believe this duplicate method was intended to be created
with a different name (`add_stub_to_context`) based on
https://github.com/Instagram/LibCST/commit/3c5aa268393952829bb7623b4bf4e865f20b1d2a

But in fact it has the same method name as the method it calls,
and as a result is completely invisible at runtime. We can just
remove it; any user broken by the change is already broken as of
the current release.

I confirmed that only the second definition of
`store_stub_in_context` exists at runtime by querying
in IPython interactively; also ran unit tests to double-check
that this has no effect:
```
> python -m unittest libcst.codemod.visitors.tests.test_apply_type_annotations
.......................................
----------------------------------------------------------------------
Ran 39 tests in 1.303s

OK
```

Ran the tests to make sure this