Fix black formatting failure on tokenize.pyi
I noticed this file was mal-formatted when using `black .` to
autoformat a different branch.

This was confusing to me, I rely on autoformatting after writing
messy code. The reason for it appears to be that black and isort
do not agree

To be specific:
- even with `profile = "black"` set, it seems that isort will add only
  one blank line on `.py` files unless otherwise specified. This is in
  direct contradiction to what's stated in isort github issues and docs,
  which claim that `profile = "black"` should solve the problem on newer
  versions of isort (I tried against the latest and it makes no
  difference)
- But if we use force `blank_lines_after_imports = 2` as the
  `pyproject.toml` currently does (and is necessary to not mess up the
  majority of our code in `*.py` files, then isort and black will
  argue with one another about how many blank lines to use in .pyi files

My solution is to skip isort on `*.pyi` files, which I think is better
than having confusing behavior.

As an alternative, or in parallel, it might be worth me opening a PR
demonstrating that `profile = "black"` isn't working for us and
referencing it in an issue in isort.