Use a class heirarchy rather than recursive types.
Summary:

This is awful.

There's about 10 slightly dense new lines of boilerplate that
carry no semantic value at all, and I still need type aliases
to give things useful names.

Plus, I have to wrap and unwrap the data at callsites. This
wouldn't be so bad if I got some nice type safety as a result
like I would in ocaml, but I'm not really getting that.

To cap it off, to the best of my knowledge there's no way to
mark a class heirarchy as "sealed" in the scala trait sense,
so we drop some of the type safety on the floor and have to use
assert statements to claim that the values we already know are
ListTreeLeafs are what we already know them to be.

I'm pretty opposed to this change, but putting it up as a
draft PR because it was requested. I'd much rather just use
recusive aliases and figure we'll fix it if pyre drops support
(which I think is unlikely, and this isn't a core module)